[CORE] Address runtime config review comments
