+++ gluten-ut/spark40/src/test/scala/org/apache/spark/sql/GlutenRuntimeConfigSuite.scala
@@ -16,26 +16,4 @@
  */
 package org.apache.spark.sql
 
-import org.apache.gluten.config.GlutenConfig
-
-class GlutenRuntimeConfigSuite extends RuntimeConfigSuite with GlutenTestsTrait {
-  test("Gluten configs report correct runtime modifiability") {
-    val conf = SparkSession.active.conf
-    assert(conf.isModifiable(GlutenConfig.COLUMNAR_FILESCAN_ENABLED.key))
-    assert(!conf.isModifiable(GlutenConfig.GLUTEN_UI_ENABLED.key))
-  }
-
-  test("GlutenConfig reads active SparkSession runtime configs") {
-    val conf = SparkSession.active.conf
-    val key = GlutenConfig.COLUMNAR_FILESCAN_ENABLED.key
-    val original = conf.get(key)
-    try {
-      conf.set(key, false)
-      assert(!GlutenConfig.get.enableColumnarFileScan)
-      conf.set(key, true)
-      assert(GlutenConfig.get.enableColumnarFileScan)
-    } finally {
-      conf.set(key, original)
-    }
-  }
-}
+class GlutenRuntimeConfigSuite extends RuntimeConfigSuite with GlutenTestsTrait {}
